WIP @voodoorai2000 review. Rebase pending.
