+++ app/models/remote_translation.rb
@@ -12,4 +12,21 @@ class RemoteTranslation < ActiveRecord::Base
     RemoteTranslationsCaller.new.delay.call(self)
   end
 
+# #Controller
+#   def translations_enqueued?(remote_translation)
+#     RemoteTranslation.where(remote_translatable_id: remote_translation["remote_translatable_id"],
+#                             remote_translatable_type: remote_translation["remote_translatable_type"],
+#                             locale: remote_translation["locale"],
+#                             error_message: nil).any?
+#   end
+
+#Helper
+  def self.remote_translation_enqueued?(remote_translation)
+    where(remote_translatable_id: remote_translation[:remote_translatable_id],
+          remote_translatable_type: remote_translation[:remote_translatable_type],
+          locale: remote_translation[:locale],
+          error_message: nil).any?
+  end
+
+
 end
